migrate CDM based unit tests to equivalent generic unit tests
